backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]>=3.3.0
bcrypt==4.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
python-multipart==0.0.20
redis==5.2.1
httpx==0.28.1
httpx-sse>=0.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
cryptography==44.0.0
litellm==1.63.2
jsonschema>=4.23.0
pypdf>=4.0.0
python-docx>=1.1.0
pgvector>=0.3.0
click>=8.0
rich>=13.0
azure-identity>=1.17.0
azure-cosmos>=4.7.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
azure-monitor-opentelemetry-exporter>=1.0.0b21
python-json-logger>=2.0.0
kubernetes>=29.0.0
azure-servicebus>=7.10.0
